+++ test/pow/store/backend/ets_cache_test.exs
@@ -6,15 +6,28 @@ defmodule Pow.Store.Backend.EtsCacheTest do
 
   @default_config [namespace: "pow:test", ttl: :timer.hours(1)]
 
+  setup do
+    pid    = self()
+    events = [
+      [:pow, EtsCache, :cache],
+      [:pow, EtsCache, :delete],
+      [:pow, EtsCache, :invalidate]
+    ]
+
+    :telemetry.attach_many("event-handler-#{inspect pid}", events, fn event, measurements, metadata, send_to: pid ->
+      send(pid, {:event, event, measurements, metadata})
+    end, send_to: pid)
+  end
+
   test "can put, get and delete records" do
     assert EtsCache.get(@default_config, "key") == :not_found
 
     EtsCache.put(@default_config, "key", "value")
-    :timer.sleep(100)
+    assert_receive {:event, [:pow, EtsCache, :cache], _measurements, %{key: "key", value: "value"}}
     assert EtsCache.get(@default_config, "key") == "value"
 
     EtsCache.delete(@default_config, "key")
-    :timer.sleep(100)
+    assert_receive {:event, [:pow, EtsCache, :delete], _measurements, %{key: "key"}}
     assert EtsCache.get(@default_config, "key") == :not_found
   end
 
@@ -43,7 +56,7 @@ defmodule Pow.Store.Backend.EtsCacheTest do
     EtsCache.put(config, "key", "value")
     :timer.sleep(50)
     assert EtsCache.get(config, "key") == "value"
-    :timer.sleep(100)
+    assert_receive {:event, [:pow, EtsCache, :invalidate], _measurements, %{key: "key"}}
     assert EtsCache.get(config, "key") == :not_found
   end
 end